requirements.txt
mysql-connector-python
pytest
pytest-csv